+++ Carro/Carro/Carro/ViewModels/CadastroProdutoViewModel.cs
@@ -54,8 +54,8 @@ namespace Carro.ViewModels
             }
         }
 
-        float _precoEntry = 0.0F;
-        public float precoEntry
+        decimal _precoEntry = 0.0m;
+        public decimal precoEntry
         {
             get
             {
