+++ Carro/Carro/Carro/ViewModels/EditarProdutoViewModel.cs
@@ -91,8 +91,8 @@ namespace Carro.ViewModels
             }
         }
 
-        float _precoEntry = 0.0F;
-        public float precoEntry
+        decimal _precoEntry = 0.0m;
+        public decimal precoEntry
         {
             get
             {
